requirements.txt
python-dotenv
langchain 
langchain-community 
langchain-openai
langchain-huggingface # Keep for other models if needed
langchain-ollama
#pypdf
#faiss-cpu
chromadb
groq
langchain-groq
beautifulsoup4
streamlit
langchain_google_genai
PyPDF2
langchain-chroma
#pdfplumber
memory_profiler
PyMuPDF
stqdm
sentence-transformers
huggingface-hub
#pysqlite3-binary 
gspread
oauth2client
pandas
google-api-python-client
pytesseract
google-generativeai
pillow
numpy
streamlit
opencv-python-headless
textblob
scipy
nltk
scikit-learn
pydantic
werkzeug
python-docx
openpyxl
fastapi
uvicorn
pydantic-settings
python-multipart
# Remove easyocr as it's no longer needed
requests
beautifulsoup4
openai # Add this for NVIDIA API client
markdown